+++ src/circleElement.ts
@@ -31,7 +31,7 @@ declare global {
        *     setTimeout(()=>popup.innerHTML=...);
        *     return ""
        *  }] */
-      tooltip?: string | ((item: Item, popup: WUPPopupElement) => string);
+      tooltip?: string | ((item: Item & { percentage: number }, popup: WUPPopupElement) => string);
     }
     interface SVGItem extends SVGPathElement {
       _relatedItem: Item;
@@ -392,10 +392,16 @@ export default class WUPCircleElement extends WUPBaseElement<WUP.Circle.Options>
       const y = r.y + segment._center.y * scale;
       return DOMRect.fromRect({ x, y, width: 0.01, height: 0.01 });
     };
-    const item = segment._relatedItem;
+    const total = this.$options.items.reduce((sum, a) => sum + a.value, 0);
+    const item = { ...segment._relatedItem, percentage: mathScaleValue(segment._relatedItem.value, 0, total, 0, 100) };
     const lbl = item.tooltip!;
+
     popup.innerText =
-      typeof lbl === "function" ? lbl.call(this, item, popup) : lbl.replace("{#}", item.value.toString());
+      typeof lbl === "function"
+        ? lbl.call(this, item, popup)
+        : lbl
+            .replace("{#}", item.value.toString())
+            .replace("{#%}", `${(Math.round(item.percentage * 10) / 10).toString()}%`);
 
     return this.appendChild(popup);
   }